ci: create workflows and do some housekeeping (#154)
Adds a CI pipeline along with some quality-of-life improvements such as an editorconfig, Makefile, functional tests refinements, dependency updates, and fixes suggested by `clippy` and `rustfmt`.

Closes #108, #80

+++ crates/btcio/src/rpc/types.rs
@@ -90,7 +90,7 @@ where
                 }
                 format!("{}{}", parts[0], padded)
             } else if parts.len() == 1 {
-                format!("{}{}", parts[0], "0".repeat(8 as usize))
+                format!("{}{}", parts[0], "0".repeat(8_usize))
             } else {
                 return Err(E::custom("Invalid amount representation"));
             };
